refactor: move bus modules into files
